refactor: 주문 도메인 로직 분리
