+++ src/test/java/kitchenpos/ui/OrderRestControllerTest.java
@@ -24,13 +24,13 @@ import java.util.Arrays;
 import java.util.List;
 
 import static kitchenpos.application.MenuGroupServiceTest.메뉴_그룹;
-import static kitchenpos.application.MenuProductTest.메뉴_상품;
-import static kitchenpos.application.MenuServiceTest.메뉴;
-import static kitchenpos.application.OrderServiceTest.주문;
-import static kitchenpos.application.TableServiceTest.두_명의_방문객;
-import static kitchenpos.application.TableServiceTest.비어있지_않은_상태;
 import static kitchenpos.application.TableServiceTest.주문_테이블;
-import static kitchenpos.application.ProductTest.상품;
+import static kitchenpos.domain.MenuProductTest.메뉴_상품;
+import static kitchenpos.domain.MenuTest.메뉴;
+import static kitchenpos.domain.OrderTableTest.두_명의_방문객;
+import static kitchenpos.domain.OrderTableTest.비어있지_않은_상태;
+import static kitchenpos.domain.OrderTest.주문;
+import static kitchenpos.domain.ProductTest.상품;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
